+++ docs/changelog.md
@@ -7,6 +7,8 @@ This changelog is used to track all major changes to WTFIX.
 **Enhancements**
 
 - Set exit codes on pipeline termination so that caller can take appropriate action.
+- Add `MessageStoreApp`, with default implementations for in-memory and redis-based stores, for caching and / or
+persisting messages.
 
 ## v0.7.0 (2019-04-21)